Changed: useEffect against copy and paste to only Slug
